requirements.txt
jupyter
matplotlib
tensorflow-addons==0.17.1
tensorflow-datasets==4.6.0
nltk